inline the "attachStaticStyles" method
