+++ lib/component.js
@@ -5,7 +5,6 @@ import {EMPTY_DIV, DOMPatcher, h} from './dom-patcher';
 import Router from './router';
 import * as hookHelpers from './component-utils/hook-helpers';
 import {getNow} from './component-utils/perf';
-import {attachStaticStyles} from './component-utils/styles';
 
 const DOCUMENT_FRAGMENT_NODE = 11;
 const ATTR_TYPE_DEFAULTS = {
@@ -14,6 +13,7 @@ const ATTR_TYPE_DEFAULTS = {
   number: 0,
   json: null,
 };
+const stylesheetCache = new Map(); // key is the tag name, value is a CSSStyleSheet instance
 
 /**
  * Definition of a Panel component/app, implemented as an HTML custom element.
@@ -317,7 +317,26 @@ class Component extends WebComponent {
 
     if (this.getConfig(`useShadowDom`)) {
       this.el = this.attachShadow({mode: `open`});
-      attachStaticStyles(this.el, this.getConfig(`css`));
+
+      const css = this.getConfig(`css`);
+      if (css) {
+        if (this.el.adoptedStyleSheets) {
+          // Attempt to cache the styles using Constructible StyleSheets if the feature is supported.
+          // Note: this technique avoids the Flash of Unstyled Content that alternative approaches like <link> tags will encounter
+          const tagName = this.el.host.localName;
+          let stylesheet = stylesheetCache.get(tagName);
+          if (!stylesheet) {
+            stylesheet = new CSSStyleSheet();
+            stylesheet.replaceSync(css);
+            stylesheetCache.set(tagName, stylesheet);
+          }
+          this.el.adoptedStyleSheets = [stylesheet];
+        } else {
+          const styleTag = document.createElement(`style`);
+          styleTag.innerHTML = css;
+          this.el.appendChild(styleTag);
+        }
+      }
     } else if (this.getConfig(`css`)) {
       throw Error(`"useShadowDom" config option must be set in order to use "css" config.`);
     } else {
